permissions

+++ main/views.py
@@ -22,34 +22,54 @@ def detail(request, id):
     return render(request, 'main/details.html', context)
 
 def add_books(request):
-    if request.method == "POST":
-        form = BookForm(request.POST or None)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            if request.method == "POST":
+                form = BookForm(request.POST or None)
 
-        if form.is_valid():
-            data = form.save(commit=False)
-            data.save()
+                if form.is_valid():
+                    data = form.save(commit=False)
+                    data.save()
+                    return redirect("main:home")
+            else:
+                form = BookForm()
+            return render(request, 'main/addbooks.html', {"form": form, "controller": "책 추가하기"})
+        else:
             return redirect("main:home")
-    else:
-        form = BookForm()
-    return render(request, 'main/addbooks.html', {"form": form, "controller": "책 추가하기"})
+
+    return redirect("accounts:login")
 
 #edit the movie
 def edit_books(request, id):
-    book = Book.objects.get(id=id)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+    
+            book = Book.objects.get(id=id)
 
-    if request.method == "POST":
-        form = BookForm(request.POST or None, instance=book)
+            if request.method == "POST":
+                form = BookForm(request.POST or None, instance=book)
 
-        if form.is_valid():
-            data = form.save(commit=False)
-            data.save()
-            return redirect("main:detail", id)
-    else:
-        form = BookForm(instance=book)
-    return render(request, 'main/addbooks.html', {"form":form, "controller": "책 정보 수정하기"})
+                if form.is_valid():
+                    data = form.save(commit=False)
+                    data.save()
+                    return redirect("main:detail", id)
+            else:
+                form = BookForm(instance=book)
+            return render(request, 'main/addbooks.html', {"form":form, "controller": "책 정보 수정하기"})
+        else:
+            return redirect("main:home")
+    
+    return redirect("accounts:login")
 
 def delete_books(request, id):
-    book = Book.objects.get(id=id)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            book = Book.objects.get(id=id)
 
-    book.delete()
-    return redirect("main:home")
+            book.delete()
+            return redirect("main:home")
+        else:
+            return redirect("main:home")
+    
+    return redirect("accounts:login")
+
